Fixing disconnected menu items click events.
