+++ DataDictionary.Main/Main.cs
@@ -171,40 +171,6 @@ namespace DataDictionary.Main
         #endregion
 
         #region Menu Events
-        private void menuSchemaItem_Click(object sender, EventArgs e)
-        { Activate((data) => new Forms.DetailDataView(data, Resources.Icon_Schema), Program.Data.DbSchemta); }
-
-        private void menuTableItem_Click(object sender, EventArgs e)
-        { Activate((data) => new Forms.DetailDataView(data, Resources.Icon_Table), Program.Data.DbTables); }
-
-        private void menuColumnItem_Click(object sender, EventArgs e)
-        { Activate((data) => new Forms.DetailDataView(data, Resources.Icon_Column), Program.Data.DbTableColumns); }
-
-        private void menuPropertyItem_Click(object sender, EventArgs e)
-        { Activate((data) => new Forms.DetailDataView(data, Resources.Icon_ExtendedProperty), Program.Data.DbExtendedProperties); }
-
-        [Obsolete()]
-        private void menuImportDbSchema_Click(object sender, EventArgs e)
-        {
-            //Program.Data.ImportDbSchemaToDomain();
-            //BuildDomainModelTree();
-        }
-
-        private void menuAttributeProperties_Click(object sender, EventArgs e)
-        { Activate((data) => new Forms.DetailDataView(data, Resources.Icon_Property), Program.Data.DomainAttributeProperties); }
-
-        private void menuAttributeAlaises_Click(object sender, EventArgs e)
-        { Activate((data) => new Forms.DetailDataView(data, Resources.Icon_Synonym), Program.Data.DomainAttributeAliases); }
-
-        private void entitiesToolStripMenuItem_Click(object sender, EventArgs e)
-        { Activate((data) => new Forms.DetailDataView(data, Resources.Icon_Entities), Program.Data.DomainEntities); }
-
-        private void entityPropertiesToolStripMenuItem_Click(object sender, EventArgs e)
-        { Activate((data) => new Forms.DetailDataView(data, Resources.Icon_Property), Program.Data.DomainEntityProperties); }
-
-        private void entityAliasToolStripMenuItem_Click(object sender, EventArgs e)
-        { Activate((data) => new Forms.DetailDataView(data, Resources.Icon_Synonym), Program.Data.DomainEntityAliases); }
-
         private void HelpContentsMenuItem_Click(object sender, EventArgs e)
         {
             if (ActiveMdiChild is Form currentForm)
@@ -223,27 +189,6 @@ namespace DataDictionary.Main
             if (ActiveMdiChild is Form currentForm)
             { Activate(() => new Dialogs.HelpSubject(currentForm)); }
         }
-
-        private void openSaveModelDatabaseMenuItem_Click(object sender, EventArgs e)
-        { Activate(() => new Forms.Model.ModelManager()); }
-
-        private void menuConstraintItem_Click(object sender, EventArgs e)
-        { Activate((data) => new Forms.DetailDataView(data, Resources.Icon_Key), Program.Data.DbConstraints); }
-
-        private void menuConstraintColumnItem_Click(object sender, EventArgs e)
-        { Activate((data) => new Forms.DetailDataView(data, Resources.Icon_KeyColumn), Program.Data.DbConstraintColumns); }
-
-        private void menuDataTypeItem_Click(object sender, EventArgs e)
-        { Activate((data) => new Forms.DetailDataView(data, Resources.Icon_DomainType), Program.Data.DbDomains); }
-
-        private void menuRoutineItem_Click(object sender, EventArgs e)
-        { Activate((data) => new Forms.DetailDataView(data, Resources.Icon_Procedure), Program.Data.DbRoutines); }
-
-        private void menuRoutineParameterItem_Click(object sender, EventArgs e)
-        { Activate((data) => new Forms.DetailDataView(data, Resources.Icon_Parameter), Program.Data.DbRoutineParameters); }
-
-        private void menuRoutineDependencyItem_Click(object sender, EventArgs e)
-        { Activate((data) => new Forms.DetailDataView(data, Resources.Icon_Dependancy), Program.Data.DbRoutineDependencies); }
 
         private void browsePropertiesCommand_Click(object sender, EventArgs e)
         { Activate(() => new Forms.Application.Property()); }
@@ -315,31 +260,6 @@ namespace DataDictionary.Main
 
         private void textEditorToolStripMenuItem_Click(object sender, EventArgs e)
         { Activate(() => new ProofOfConcept.TextEditor()); }
-
-
-        private void manageLibrariesCommand_ButtonClick(object sender, EventArgs e)
-        { Activate(() => new Forms.Library.LibraryManager()); }
-
-        private void manageDatabasesCommand_ButtonClick(object sender, EventArgs e)
-        { Activate(() => new Forms.Database.CatalogManager()); }
-
-        private void menuCatalogItem_Click(object sender, EventArgs e)
-        { Activate((data) => new Forms.DetailDataView(data, Resources.Icon_Database), Program.Data.DbCatalogs); }
-
-        private void menuAttributes_Click(object sender, EventArgs e)
-        { Activate((data) => new Forms.DetailDataView(data, Resources.Icon_Attribute), Program.Data.DomainAttributes); }
-
-        private void subjectAreaToolStripMenuItem_Click(object sender, EventArgs e)
-        { Activate((data) => new Forms.DetailDataView(data, Resources.Icon_Diagram), Program.Data.ModelSubjectAreas); }
-
-        private void browseHelpCommand_Click(object sender, EventArgs e)
-        { Activate((data) => new Forms.DetailDataView(data, Resources.Icon_HelpTableOfContent), Program.Data.HelpSubjects); }
-
-        private void viewLibrarySourceCommand_Click(object sender, EventArgs e)
-        { Activate((data) => new Forms.DetailDataView(data, Resources.Icon_Library), Program.Data.LibrarySources); }
-
-        private void viewLibraryMemberCommand_Click(object sender, EventArgs e)
-        { Activate((data) => new Forms.DetailDataView(data, Resources.Icon_Class), Program.Data.LibraryMembers); }
 
         private void openToolStripMenuItem_Click(object sender, EventArgs e)
         {
